requirements.txt
absl-py
appdirs
apturl
astunparse
atomicwrites
attrs
backcall
backports.entry-points-selectable
bcrypt
beautifulsoup4
bleach
blinker
blosc
Brlapi
cachetools
caffe
certifi
cffi
chardet
Click
cloudpickle
colorama
command-not-found
cryptography
cupshelpers
cycler
Cython
dask
dbus-python
decorator
defer
defusedxml
distlib
distro
distro-info
dlib
duplicity
easyocr
entrypoints
et-xmlfile
fasteners
filelock
flake8
flatbuffers
fsspec
future
gast
gdown
google-auth
google-auth-oauthlib
google-pasta
grpcio
h5py
html5lib
httplib2
idna
imageio
importlib-metadata
ipykernel
ipython
ipython-genutils
ipywidgets
jdcal
jedi
Jinja2
joblib
jsonschema
jupyter-client
jupyter-console
jupyter-core
Keras
Keras-Applications
Keras-Preprocessing
keyring
kiwisolver
language-selector
launchpadlib
lazr.restfulclient
lazr.uri
locket
lockfile
loguru
louis
lxml
macaroonbakery
Mako
Markdown
MarkupSafe
matplotlib
mccabe
mistune
monotonic
more-itertools
mpi4py
nbconvert
nbformat
netifaces
networkx
nose
notebook
numexpr
numpy
nvidia-ml-py
oauthlib
olefile
opencv-python
openpyxl
opt-einsum
packaging
pandas
pandocfilters
parameterized
paramiko
parso
partd
pexpect
pickleshare
Pillow
platformdirs
pluggy
ply
prometheus-client
prompt-toolkit
protobuf
psutil
py
pyasn1
pyasn1-modules
pycairo
pycodestyle
pycparser
pycuda
pycups
pydot
pyflakes
pygame
Pygments
PyGObject
pygpu
pyinotify
PyJWT
pymacaroons
PyNaCl
pyOpenSSL
pyparsing
pyRFC3339
pyrsistent
PySocks
pytest
python-apt
python-bidi
python-dateutil
python-debian
pytools
pytz
PyWavelets
pyxdg
PyYAML
pyzmq
reportlab
requests
requests-oauthlib
requests-unixsocket
rsa
scikit-cuda
scikit-image
scikit-learn
scipy
screen-resolution-extra
SecretStorage
Send2Trash
simplejson
six
soupsieve
systemd-python
tables
tensorboard
tensorflow-estimator
tensorflow-gpu
termcolor
terminado
testpath
Theano
toolz
torch
torchaudio
torchvision
tornado
tqdm
traitlets
typing-extensions
ubuntu-advantage-tools
ubuntu-drivers-common
ufw
unattended-upgrades
urllib3
usb-creator
virtualenv
wadllib
wcwidth
webencodings
Werkzeug
widgetsnbextension
wrapt
xkit
xlrd
xlwt
zipp
